scheduler/requirements.txt
asgiref==3.5.2
celery==5.2.6
fastapi==0.77.1
httpx==0.17.1
pandas==1.4.2
python-dateutil==2.8.1
python-dotenv==0.20.0
redis==4.3.1